Codespace congenial palm tree xg46gjrjjjjh6r9r

+++ Super_Trunfo_Logica.C
@@ -2,6 +2,9 @@
 #include <string.h>
 #include <math.h>
 
+//Programa para atividade 2 de super trunfo.
+
+
 
 int main(){
     //Declaração de Variáveis
@@ -10,7 +13,7 @@ int main(){
     float area1, area2, pib1, pib2, pibPerCap1, pibPerCap2, densDemo1, densDemo2; //variáveis tipo float das cartas
     int bolPop, bolAre, bolPib, bolPoT, bolPPC, bolDeD, bolSuP; //variáveis booleanas auxiliares
     char resPopulac[7], resArea[7], resPib[7], resPontTur[7], resPibPerCap[7], resDensDemo[7], resSuperPoder[7]; //Variáveis para guardar o vencedor de cada modalidade
-    float superPoder1=0, superPoder2=0;
+    double superPoder1=0, superPoder2=0;
     //Apresentação do jogo e coleta de dados
 
     printf("Bem vindo ao Super Trunfo das Cidades!!!\n");
@@ -181,7 +184,7 @@ int main(){
     printf("Número de pontos turísticos: %d\n", pontTur1);
     printf("Densidade Demográfica: %.2f habitantes por km²\n", densDemo1);
     printf("PIB per capita: %.2f reais\n", pibPerCap1);
-    printf("Super Poder: %.6f\n", superPoder1);
+    printf("Super Poder: %.6lf\n", superPoder1);
     printf("\n");
 
     //carta 2
@@ -196,7 +199,7 @@ int main(){
     printf("Número de pontos turísticos: %d\n", pontTur2);
     printf("Densidade Demográfica: %.2f habitantes por km²\n", densDemo2);
     printf("PIB per capita: %.2f reais\n", pibPerCap2);
-    printf("Super Poder: %.6f\n", superPoder2);
+    printf("Super Poder: %.6lf\n", superPoder2);
     printf("\n");
     printf("\n");
